correction des algorithmes et du calcul du cout final

+++ TP1Functions.h
@@ -27,7 +27,7 @@ typedef struct dataSet
 
 int read_TP1_instance(FILE*fin,dataSet* dsptr);
 int KP_greedy(dataSet* dsptr);
-int KP_LP(dataSet* dsptr);
+double KP_LP(dataSet* dsptr);
 int KP_dynamic(dataSet* dsptr);
